# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IAccessTokenData": {
				"properties": {
					"token": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DetailedError": {
				"properties": {
					"message": {
						"type": "string"
					},
					"standardizedErrorCode": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"standardizedErrorCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServerResponse_IAccessTokenData_": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/IAccessTokenData"
					},
					"error": {
						"type": "string"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/DetailedError"
						},
						"type": "array"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAgentSignupDto": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"emailAddress": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"passwod": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"emailAddress",
					"phoneNumber",
					"passwod"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "support-backend",
		"version": "1.0.0",
		"description": "Support Service Backend",
		"contact": {
			"name": "Mubarak Aminu"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/auth/agent/singup": {
			"post": {
				"operationId": "AgentSignup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IServerResponse_IAccessTokenData_"
								}
							}
						}
					}
				},
				"tags": [
					"Auth Service"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAgentSignupDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}